+++ Task.h
@@ -13,13 +13,13 @@
 class Task {
 public:
 	//Constructors & Destructor
-	Task(const unsigned int a_machineId, const unsigned int a_duration);
+	Task(const unsigned short a_machineId, const unsigned short a_duration);
 	Task(const Task& rhs);
 	virtual ~Task();
 
 	//Getters & Setters
-	unsigned int getDuration() const;
-	unsigned int getMachineId() const;
+	unsigned short getDuration() const;
+	unsigned short getMachineId() const;
 	unsigned long long getEndTime() const;
 	void setEndTime(unsigned long long startTime);
 	bool isTaskCompleted() const;
@@ -27,8 +27,8 @@ public:
 
 private:
 	//Attributes
-	unsigned int machineId;
-	unsigned int duration;
+	unsigned short machineId;
+	unsigned short duration;
 	unsigned long long endTime;
 	bool taskCompleted;
 };
